Sequences and Series Update
Updated sequences and series to user a newer release of jsxgraph because the zoom features were not working well on the older version. I also updated math.js to a newer version. This in turn made me update how plotting works in MathFunctions.js which will likely break a lot of my other apps, but would certainly be better in the end.