updated module 1 trig videos
I updated the module 1 trig videos to use a cdn version of jsx instead of the one hosted at bayreuth. I also fixed an error in the bearings applet because East and West were backwards. I also fixed Trig functions defined by points because an issue with a checkbox was causing it to malfunction